meta-isar: Add a recipe that configures ci user
This creates `ci` user on a target and configures it for
non-interactive access to the image via SSH.

Non-interactive access is provided by placing testsuite's public part
of ssh key to `authorized_keys`.

This allows non-interactive SSH access to the machine with executing
custom commands on the guest VM.

Also, NetworkManager is installed to the image to make ethernet on
target auto configured and thus port forwarding from host to qemu
machine would work.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>